Build 2.0.0.72
---Did some C++ code analysis to catch programming errors (surprisingly few)
---Re-fixed that enhancement issue which broke enhancements requiring other enhancements. (Original issue reported by Kenaustin)
---Adjusted how the "Hand and a Half" Two Handed Specialty ability bonus affects totals when SWF is enabled (Reported by shoka)
---Equipped Shields should now correctly award their base PRR bonus for being equipped (Reported by Kenaustin)
---The "Shield Mastery" feat lines should now provide the correct PRR at each feat level (Reported by Kenaustin)
---Cannith crafted "Potency" now correctly affects each individual spell power, not universal spell power (Reported by Jennan)
---"Battle Engineer: Reconstructed Arms" now correctly awards +20% Competence bonus to HP (Reported by yugodealer)
---"Mechanic: Expert Builder" now correctly costs 1ap (Reported by Jokersir2)
---Auto granted spells now have additional text in their tooltips to help delineate them correctly (Reported by DougSK)
---Typos in "Improved Precise Shot" and "Nature's Protector: Savage Roar" fixed (Reported by MrEkshin)
---Forum export will no longer erroneously remove some table lines under certain circumstances (Reported by dpm5d)

U78 Lamannia Changes
---Arcane Trickster class added
------Past Lives setup
------Class specific feats added
------Arcane Trickster enhancement tree added
------Active Rogue past life feat requirements updated
------Various metamagic feats updated to work correctly with Arcane Trickster
------New spells/SLAs added
------Thief Acrobat tree updated to work for Arcane Trickster
------Mechanic tree updated to work for Arcane Trickster
------Epic/Legendary Knowledge should correctly award Arcane Trickster caster levels

+++ DDOBuilder/BreakdownItemDestinyAps.cpp
@@ -73,7 +73,7 @@ void BreakdownItemDestinyAps::CreateOtherEffects()
                         Effect_DestinyAPBonus,
                         "Epic Levels",
                         "Epic",
-                        epicLevels * 4);
+                        static_cast<double>(epicLevels) * 4);
                     AddOtherEffect(elBonus);
                 }
                 if (level >= MAX_EPIC_LEVEL)
@@ -88,7 +88,7 @@ void BreakdownItemDestinyAps::CreateOtherEffects()
                         Effect_DestinyAPBonus,
                         "Legendary Levels",
                         "Legendary",
-                        level * 4);
+                        static_cast<double>(level) * 4);
                     AddOtherEffect(llBonus);
                 }
             }
